made autonomous path, requires place purple and place yellow Actions to be made still
